feat: Split SNS webhook server from main HTTP server
