+++ src/core/infra/adapters/ExpressMiddlewareAdapter.ts
