+++ src/infra/http/routes/contacts.routes.ts
@@ -1,7 +1,8 @@
 import express from 'express'
 
-import { adaptMiddleware } from '../adapters/ExpressMiddlewareAdapter'
-import { adaptRoute } from '../adapters/ExpressRouteAdapter'
+import { adaptMiddleware } from '@core/infra/adapters/ExpressMiddlewareAdapter'
+import { adaptRoute } from '@core/infra/adapters/ExpressRouteAdapter'
+
 import { makeBlockContactController } from '../factories/controllers/BlockContactControllerFactory'
 import { makeCreateContactController } from '../factories/controllers/CreateContactControllerFactory'
 import { makeGetContactDetailsController } from '../factories/controllers/GetContactDetailsControllerFactory'
