+++ src/infra/http/routes/index.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express'
 
 import { contactsRouter } from './contacts.routes'
-import { eventsRouter } from './events.routes'
 import { messagesRouter } from './messages.routes'
 import { recipientsRouter } from './recipients.routes'
 import { sendersRouter } from './senders.routes'
@@ -15,7 +14,6 @@ const router = Router()
 router.use('/users', usersRouter)
 router.use('/sessions', sessionsRouter)
 router.use('/messages', messagesRouter)
-router.use('/events', eventsRouter)
 router.use('/contacts', contactsRouter)
 router.use('/senders', sendersRouter)
 router.use('/tags', tagsRouter)
