+++ src/infra/http/routes/messages.routes.ts
@@ -1,7 +1,8 @@
 import express from 'express'
 
-import { adaptMiddleware } from '../adapters/ExpressMiddlewareAdapter'
-import { adaptRoute } from '../adapters/ExpressRouteAdapter'
+import { adaptMiddleware } from '@core/infra/adapters/ExpressMiddlewareAdapter'
+import { adaptRoute } from '@core/infra/adapters/ExpressRouteAdapter'
+
 import { makeCreateMessageController } from '../factories/controllers/CreateMessageControllerFactory'
 import { makeGetMessageDetailsController } from '../factories/controllers/GetMessageDetailsControllerFactory'
 import { makeGetMessageStatsController } from '../factories/controllers/GetMessageStatsControllerFactory'
