+++ src/infra/http/routes/sessions.routes.ts
@@ -1,6 +1,7 @@
 import express from 'express'
 
-import { adaptRoute } from '../adapters/ExpressRouteAdapter'
+import { adaptRoute } from '@core/infra/adapters/ExpressRouteAdapter'
+
 import { makeAuthenticateUserController } from '../factories/controllers/AuthenticateUserControllerFactory'
 
 const sessionsRouter = express.Router()
