+++ src/infra/sns-webhook/factories/middlewares/AmazonSNSValidatorMiddlewareFactory.ts
@@ -1,5 +1,5 @@
 import { Middleware } from '@core/infra/Middleware'
-import { AmazonSNSValidatorMiddleware } from '@infra/http/middlewares/AmazonSNSValidatorMiddleware'
+import { AmazonSNSValidatorMiddleware } from '@infra/sns-webhook/middlewares/AmazonSNSValidatorMiddleware'
 
 export function makeAmazonSNSValidatorMiddleware(): Middleware {
   const amazonSNSValidatorMiddleware = new AmazonSNSValidatorMiddleware()
