+++ src/infra/sns-webhook/server.ts
@@ -1,0 +1,5 @@
+import { app } from './app'
+
+app.listen(process.env.PORT, () => {
+  console.log(`Server running on http://localhost:${process.env.PORT}`)
+})
